Renamed IConsumer to IEventHandler.
This is done as a first step in a refactoring that separates event
handlers from command handlers. The concept is that in CQRS, there must be
exactly one command handler per command, but zero to many event handlers
per event.
This rename makes it clear that the current code base is lacking a command
handler type, and that CapacityGate should be a command handler, and not
an event handler.